Typogrpahy changes, add variable font version of Archivo, remove unused components
