+++ src/App.css
@@ -37,7 +37,8 @@ a {
   /* letter-spacing: 0.5px; */
   /* line-height: 0.5; */
 
-  font-size: 0.8rem;
+  /* font-size: 0.8rem; */
+  font-size: 0.7rem;
   letter-spacing: 0.025rem;
   line-height: 1;
   font-weight: 400;
@@ -77,22 +78,25 @@ a {
 @media (min-width: 800px) {
   .page .jumbo-text {
     font-size: 2rem; /* fallback */
-    font-size: min(max(2rem, 3vw), 2.5rem);
+    /* font-size: min(max(2rem, 3vw), 2.5rem); */
+    font-size: min(max(2rem, 4vw), 54px);
 
     letter-spacing: -0.1rem;
     line-height: 0.9;
-    font-weight: 600;
+    /* font-weight: 600; */
 
     /* letter-spacing: 0; */
     word-spacing: 0;
-    line-height: normal;
-    line-height: 1;
-    font-weight: 600;
+    /* line-height: normal; */
+    /* line-height: 1; */
+    /* font-weight: 600; */
+    font-weight: 700;
   }
 
   .page h3,
   .tag h3 {
-    font-size: 0.9rem;
+    /* font-size: 0.9rem; */
+    font-size: 0.8rem;
   }
 }
