+++ src/components/Cursor/Cursor.css
@@ -20,7 +20,7 @@
 
 .about-page .cursor {
   opacity: 1;
-  background-color: yellow;
+  background-color: var(--yellow);
   transform: translate(-50%, -50%) scale(1);
 }
 
@@ -31,7 +31,7 @@
 
 .cursor.link {
   opacity: 1;
-  background-color: yellow;
+  background-color: var(--yellow);
   transform: translate(-50%, -50%) scale(0.4);
 }
