--- src/components/FadeInAnimation/FadeInAnimation.css
@@ -1,29 +1,0 @@
-.fade-in-animation__container{
-  opacity: 0;
-  transform: translateY(100px);
-  transition: opacity 1s ease-in-out, transform 1s ease-in-out;
-  position: relative;
-  top: 200px;
-}
-/* .fade-in-animation__container.shown{
-  opacity: 1;
-  transform: translateY(0);
-  transition: opacity 1s ease-in-out, transform 1s ease-in-out;
-}
-
-.animation__show{
-  -webkit-animation:fadeIn 3s ;
-  animation: fadeIn 3s ;
-}
-
-@-webkit-keyframes fadeIn {
-  0% { opacity: 0; transform: translateY(100px);}
-  50% { opacity: .5; transform: translateY(50px); } 
-  100% { opacity: 0; transform: translateY(0px); }
-}
-
-@keyframes fadeIn {
-  0% { opacity: 0; transform: translateY(100px);}
-  50% { opacity: .5; transform: translateY(50px); } 
-  100% { opacity: 0; transform: translateY(0px); }
-} */
