--- src/components/FadeInOutAnimation/FadeInOutAnimation.css
@@ -1,56 +1,0 @@
-.fade-in {
-  opacity: 1;
-  /* transform: translateY(-50px); */
-  /* transform: translateY(-5%); */
-  /* position: relative;
-  top: 50px;
-   */
-  transition: all 2.5s;
-
-  /* animation-name: fadeUp;
-  animation-duration: 2s;
-  animation-fill-mode: forwards; */
-}
-
-.fade-out {
-  opacity: 0;
-  /* transform: translateY(-50px); */
-  transition: all 2s;
-  /* animation-name: fadeDown;
-  animation-duration: 1s;
-  animation-fill-mode: forwards; */
-}
-
-
-@keyframes fadeUp {
-  from {
-    opacity: 0;
-    transform: translateY(100px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-@keyframes fadeDown {
-  from {
-    opacity: 1;
-    transform: translateY(0);
-  }
-  to {
-    opacity: 0;
-    transform: translateY(-10px);
-  }
-}
-
-@media(max-width: 800px){
-  .fade-in {
-    opacity: 1;
-    transform: translateY(0%);
-  }
-  
-  .fade-out {
-    opacity: 1;
-    transition: all 2s;
-  }
-}
