--- src/components/FadeInUpOutUp/FadeInUpOutUp.css
@@ -1,13 +1,0 @@
-.fade-element {
-  transition: all 2s;
-}
-
-.fade-in-up {
-  bottom: 0;
-  opacity: 1;
-}
-
-.fade-out-up {
-  bottom: "-100%";
-  opacity: 0;
-}
