+++ src/components/NavBar/NavBar.css
@@ -17,6 +17,26 @@
   padding: 0 2%;
 }
 
+.nav-bar__item {
+  width: 25%;
+  display: flex;
+}
+
+.nav-bar__logo {
+  /* width: auto; */
+  flex-grow: 1;
+  justify-content: center;
+}
+
+.nav-bar__item:last-of-type {
+  /* text-align: right; */
+  justify-content: flex-end;
+}
+
+.nav-bar__item .tag {
+  width: fit-content;
+}
+
 .nav-bar__logo img {
   width: auto;
   height: 25px;
@@ -29,6 +49,13 @@
   transition: transform 0.5s ease;
   cursor: pointer;
 }
+
+/* .nav-bar .tag-link h3 {
+  transition: font-weight 1s;
+}
+.nav-bar .tag-link:hover h3 {
+  font-weight: 600;
+} */
 
 @media (min-width: 800px) {
   .nav-bar {
