+++ src/components/NavMenu/NavMenu.css
@@ -6,9 +6,9 @@
   height: 100dvh;
   overflow-y: auto;
   z-index: 999;
-  background-color: #aacc12;
+  background-color: #000000;
   /* animation: color-transition-menu 300s linear 0s infinite forwards, menuOpen 1s ease-out 0s 1 forwards; */
-  animation: color-transition-menu 300s linear 0s infinite forwards;
+  /* animation: color-transition-menu 300s linear 0s infinite forwards; */
   display: flex;
   justify-content: center;
   /* transform: translate(0, 0); */
@@ -63,7 +63,8 @@
   width: fit-content;
   color: white;
   text-transform: uppercase;
-  font-weight: 600;
+  /* font-weight: 600; */
+  font-weight: 700;
   font-size: 3.5rem;
   letter-spacing: -0.25rem;
   word-spacing: 0.2rem;
@@ -123,7 +124,7 @@
   }
 }
 
-@keyframes color-transition-menu {
+/* @keyframes color-transition-menu {
   0% {
     background-color: #aacc12;
   }
@@ -199,90 +200,7 @@
   100% {
     background-color: #aacc12;
   }
-}
-
-@keyframes color-transition {
-  /* for background */
-  0% {
-    background-color: #000000;
-    /* background-color: #aacc12; */
-  }
-  0.5% {
-    background-color: #aacc12;
-  }
-  6% {
-    background-color: #00a742;
-  }
-  8.33% {
-    background-color: #f95300;
-  }
-  12.5% {
-    background-color: #66ac83;
-  }
-  16.67% {
-    background-color: #5b622c;
-  }
-  20.83% {
-    background-color: #88b6a9;
-  }
-  25% {
-    background-color: #e34165;
-  }
-  29.17% {
-    background-color: #ff3312;
-  }
-  33.33% {
-    background-color: #485db3;
-  }
-  37.5% {
-    background-color: #e74e21;
-  }
-  41.67% {
-    background-color: #e5f1b8;
-  }
-  45.83% {
-    background-color: #c3a875;
-  }
-  50% {
-    background-color: #991f22;
-  }
-  54.17% {
-    background-color: #f67133;
-  }
-  58.33% {
-    background-color: #f1f3e9;
-  }
-  62.5% {
-    background-color: #f8bb00;
-  }
-  66.67% {
-    background-color: #4766b2;
-  }
-  70.83% {
-    background-color: #ed1e24;
-  }
-  75% {
-    background-color: #7d9fad;
-  }
-  79.17% {
-    background-color: #296eae;
-  }
-  83.33% {
-    background-color: #babcbe;
-  }
-  87.5% {
-    background-color: #ff6c2f;
-  }
-  91.67% {
-    background-color: #f9db16;
-  }
-  95.83% {
-    background-color: #e1e9ef;
-  }
-  100% {
-    background-color: #aacc12;
-  }
-}
+} */
 
 @keyframes nav-menu__close-bar-left {
   0% {
