+++ src/components/TagLink/TagLink.css
@@ -11,10 +11,30 @@
 
 .tag-link-nav h3 {
   text-transform: uppercase;
+  color: #ffffff;
 }
 
+/* @keyframes vars {
+  0% {
+    font-variation-settings: "wght" 500;
+  }
+
+  100% {
+    font-variation-settings: "wght" 700;
+  }
+} */
+
 .tag-link-nav h3 {
-  color: #ffffff;
+  font-variation-settings: "wght" 500;
+  /* letter-spacing: 0.025rem; */
+  transition: font-variation-settings 0.3s ease;
+  /* transition: font-variation-settings 2s ease; */
+}
+.tag-link-nav h3:hover {
+  font-variation-settings: "wght" 750;
+  /* letter-spacing: 0.0125rem; */
+  /* letter-spacing: -0.1px; */
+  /* animation: vars 0.5s 1 ease both; */
 }
 
 .tag-link-underline {
@@ -22,12 +42,16 @@
   bottom: 2px;
   width: 100%;
   height: 2px;
-  transform: scaleX(0);
+  transform: scaleX(1);
   transition: transform 0.3s ease;
 }
 
+.tag-link-nav .tag-link-underline {
+  display: none;
+}
+
 .tag-link-underline.hover {
-  transform: scaleX(1);
+  transform: scaleX(0);
 }
 
 .tag-link-nav .tag-link-underline hr {
